Move graphics components on actors to child object
This fixes an issue where the dragon's center was near its rear end
instead of what looked like its actual center, which resulted in the
enemies' ranged projectiles sometimes being aimed into the ground. The
other actors' graphics components were modified to keep things
standardized.

Let's
- move the SpriteRenderer, Animator, and PhotonAnimatorView components
  on actors to a Graphics child object
- create a CombatAnimationEventHandler class to handle animation events
  on the child object
