+++ Assets/Scripts/Actor Components/Combat/Combat.cs
@@ -16,7 +16,7 @@ public class Combat : MonoBehaviour
         public CombatAbility Ability { get => ability; }
     }
 
-    [SerializeField] protected Collider2D collider;
+    [SerializeField] protected Collider2D collider2d;
     [SerializeField] protected Rigidbody2D rigidbody2d;
     [SerializeField] protected Animator animator;
     [SerializeField] private LayerMask attackEffectLayer;
@@ -40,7 +40,7 @@ public class Combat : MonoBehaviour
     private Dictionary<CombatAbilityIdentifier, CombatAbility> identifierToAbilityDictionary
         = new Dictionary<CombatAbilityIdentifier, CombatAbility>();
 
-    public Collider2D Collider { get => collider; }
+    public Collider2D Collider2d { get => collider2d; }
     public Rigidbody2D Rigidbody2d { get => rigidbody2d; }
     public Animator Animator { get => animator; }
     public LayerMask AttackEffectLayer { get => attackEffectLayer; }
@@ -80,35 +80,9 @@ public class Combat : MonoBehaviour
         identifierToAbilityDictionary[ability].End(this);
     }
 
-    public virtual void Attack() { }
-
-    public void ExecuteAttackEffect()
-    {
-        if (CombatStateMachine.CurrState is AttackState attackState)
-        {
-            attackState.ExecuteAttackEffect();
-        }
-    }
-
-    public void OnAttackEnd()
-    {
-        if (CombatStateMachine.CurrState is AttackState)
-        {
-            CombatStateMachine.Exit();
-        }
-    }
-
     public void Stun()
     {
         CombatStateMachine.ChangeState(new StunState(this));
-    }
-
-    public void OnStunEnd()
-    {
-        if (CombatStateMachine.CurrState is StunState)
-        {
-            CombatStateMachine.Exit();
-        }
     }
 
     public void Hurt()
@@ -127,14 +101,6 @@ public class Combat : MonoBehaviour
                 CombatStateMachine.ChangeState(new DeathState(this));
                 deathEvent.Invoke();
             }
-        }
-    }
-
-    public void OnHurtEnd()
-    {
-        if (CombatStateMachine.CurrState is HurtState)
-        {
-            CombatStateMachine.Exit();
         }
     }
 }
