+++ Assets/Scripts/Actor Components/Combat/CombatAnimationEventHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 414b68c867b35ca429496dbab468a90c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
